refactor: confine reservation request dto to controller layer
