+++ src/main/java/roomescape/controller/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.controller;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
