+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,9 +1,9 @@
 package roomescape.service;
 
+import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.ReservationRequest;
 import roomescape.dao.ReservationDao;
 import roomescape.domain.Reservation;
 
@@ -22,8 +22,8 @@ public class ReservationService {
     }
 
     @Transactional
-    public Reservation createReservation(ReservationRequest reservationRequest) {
-        Long id = reservationDao.insertWithKeyHolder(reservationRequest);
+    public Reservation createReservation(String name, LocalDate date, Long timeId) {
+        Long id = reservationDao.insertWithKeyHolder(name, date, timeId);
         return reservationDao.findReservationById(id);
     }
